fix:data type

+++ tasks/kolotukhin_a_gaussian_blur/common/include/common.hpp
@@ -10,7 +10,7 @@
 namespace kolotukhin_a_gaussian_blur {
 
 // pixel data, image wigth, image height
-using InType = std::tuple<std::vector<uint8_t>, size_t, size_t>;
+using InType = std::tuple<std::vector<uint8_t>, int, int>;
 using OutType = std::vector<uint8_t>;
 using TestType = std::tuple<InType, std::vector<uint8_t>, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
